api/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.4.1
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==3.2.6
django-celery==3.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==3.0.1
kombu==3.0.37
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.19.5
packaging==21.0
passlib==1.7.4
pep8==1.7.1
Pillow==8.3.2
# pkg_resources==0.0.0
proglog==0.1.9
psycopg2==2.9.1
pycrypto==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
tqdm==4.62.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
